Support PHP5.x migration
This change ensures that projects created in PHP 5.x will work as they do in the latest PHP 8.3.

Add CacheModule and version check in AppInjector

The current commit introduces the CacheModule and inserts it in the AppInjector module's installation process. In addition, a function named 'checkVersion' has been created to ensure the existence of 'Doctrine\Common\Cache\FilesystemCache', failing which an exception is triggered. The check is performed during the AppInjector’s construction. This ensures that a doctrine cache ^1.0 is always present for the functioning of AppInjector.